(General) Refactor from `spadium.flyftp.serialization` to `spadium.flyftp.storage`
