+++ app/src/main/java/spadium/flyftp/screens/HistoryScreen.kt
@@ -6,7 +6,7 @@ import androidx.compose.material3.ListItem
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import spadium.flyftp.serialization.ServerType
+import spadium.flyftp.storage.config.ServerType
 
 @Composable
 fun HistoryScreen() {
