+++ app/src/main/java/spadium/flyftp/storage/config/Config.kt
@@ -1,4 +1,4 @@
-package spadium.flyftp.serialization
+package spadium.flyftp.storage.config
 // TODO: Implement new config system
 enum class ColorModes {
     MATERIAL3, DARK, LIGHT, AMOLED, DEFAULT
