+++ app/src/main/java/spadium/flyftp/storage/downloads/History.kt
@@ -1,3 +1,3 @@
-package spadium.flyftp.serialization
+package spadium.flyftp.storage.downloads
 
 data class History(val time: Long, val server: Any, )
