+++ app/src/main/java/spadium/flyftp/ui/theme/Theme.kt
@@ -16,7 +16,7 @@ import androidx.compose.ui.graphics.toArgb
 import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalView
 import androidx.core.view.WindowCompat
-import spadium.flyftp.serialization.ColorModes
+import spadium.flyftp.storage.config.ColorModes
 
 private val DarkColorScheme = darkColorScheme(
         primary = Purple80,
